feature: use cosmjs (#14)
* refactor getbalance

* make getBalance generic

* add envar back. lints. ibc service cleanup

* refactor ibc transfer logic. add ibc channel to envar.

* refactor justfiles.

* better command name

+++ web/src/components/DepositCard/DepositCard.tsx
@@ -6,12 +6,10 @@ import AnimatedArrowSpacer from "components/AnimatedDownArrowSpacer/AnimatedDown
 import EthWalletConnector from "features/EthWallet/components/EthWalletConnector/EthWalletConnector";
 import { NotificationsContext } from "contexts/NotificationsContext";
 import { useEthWallet } from "features/EthWallet/hooks/useEthWallet";
-import { getJSON } from "services/api";
-import { sendIbcTransfer } from "services/ibc";
+import { getBalance, sendIbcTransfer } from "services/ibc";
 import { getKeplrFromWindow } from "services/keplr";
-import type { Balances } from "types";
+import { useIbcChainSelection } from "features/IbcChainSelector/hooks/useIbcChainSelection";
 import Dropdown from "../Dropdown/Dropdown";
-import { useIbcChainSelection } from "../../features/IbcChainSelector/IbcChainSelector";
 
 export default function DepositCard(): React.ReactElement {
   const { addNotification } = useContext(NotificationsContext);
@@ -29,6 +27,7 @@ export default function DepositCard(): React.ReactElement {
     useState<boolean>(false);
   const [hasTouchedForm, setHasTouchedForm] = useState<boolean>(false);
   const [isLoading, setIsLoading] = useState<boolean>(false);
+  const [isLoadingBalance, setIsLoadingBalance] = useState<boolean>(false);
   const [isAnimating, setIsAnimating] = useState<boolean>(false);
 
   // set recipient address if userAccount available,
@@ -48,36 +47,39 @@ export default function DepositCard(): React.ReactElement {
     checkIsFormValid(recipientAddress, amount);
   }, [recipientAddress, amount]);
 
-  const getBalance = async () => {
+  const getAndSetBalance = async () => {
     if (!selectedIbcChain) {
       return;
     }
-    const key = await window.keplr?.getKey(selectedIbcChain.chainId);
-
-    if (key) {
-      const uri = `${selectedIbcChain.rest}/cosmos/bank/v1beta1/balances/${key.bech32Address}?pagination.limit=1000`;
-
-      const data = await getJSON<Balances>(uri);
-      const balance = data.balances.find((balance) => balance.denom === "utia");
-      const tiaDecimal = selectedIbcChain.currencies.find(
-        (currency: { coinMinimalDenom: string }) =>
-          currency.coinMinimalDenom === "utia",
-      )?.coinDecimals;
-
-      if (balance) {
-        const amount = new Dec(balance.amount, tiaDecimal);
-        setBalance(`${amount.toString(tiaDecimal)} TIA`);
-      } else {
-        setBalance("0 TIA");
-      }
+    try {
+      setIsLoadingBalance(true);
+      const balance = await getBalance(selectedIbcChain);
+      setBalance(balance);
+    } catch (e) {
+      console.error(e);
+      setBalance("Error fetching balance");
+    } finally {
+      setIsLoadingBalance(false);
     }
   };
 
   const sendBalance = async () => {
+    if (!selectedIbcChain) {
+      addNotification({
+        toastOpts: {
+          toastType: NotificationType.WARNING,
+          message: "Please select a chain first.",
+          onAcknowledge: () => {},
+        },
+      });
+      return;
+    }
+
     setIsLoading(true);
     setIsAnimating(true);
     try {
       await sendIbcTransfer(
+        selectedIbcChain,
         fromAddress,
         recipientAddress,
         DecUtils.getTenExponentN(6).mul(new Dec(amount)).truncate().toString(),
@@ -137,7 +139,7 @@ export default function DepositCard(): React.ReactElement {
     try {
       const key = await keplr.getKey(selectedIbcChain.chainId);
       setFromAddress(key.bech32Address);
-      await getBalance();
+      await getAndSetBalance();
     } catch (e) {
       if (e instanceof Error) {
         console.error(e.message);
@@ -223,8 +225,13 @@ export default function DepositCard(): React.ReactElement {
             </button>
           </div>
           <div>
-            {fromAddress && (
+            {fromAddress && !isLoadingBalance && (
               <p className="mt-2 has-text-light">Balance: {balance}</p>
+            )}
+            {fromAddress && isLoadingBalance && (
+              <p className="mt-2 has-text-light">
+                Balance: <i className="fas fa-spinner fa-pulse" />
+              </p>
             )}
           </div>
         </div>
